# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: |
    API specification for Legal Lawyer AI backend.

    Safe Mode:
    - Only endpoints that declare the `X-Safe-Mode` header enforce it.
    - When `X-Safe-Mode: 1`, those endpoints block writes and return 403.
servers:
  - url: https://gowaned-beckham-unlawful.ngrok-free.dev
  - url: http://127.0.0.1:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    safeModeHeader:
      type: apiKey
      in: header
      name: X-Safe-Mode
      description: "Set to '1' to enable safe mode (read-only, no persistence)"
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
    RateLimitError:
      type: object
      properties:
        message: { type: string, example: "Rate limit exceeded" }
        error: { type: string, example: "Too Many Requests" }
...
    SignupResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          type: object
          required: [id, name, email, isEmailVerified]
          properties:
            id: { type: integer, format: int64 }
            name: { type: string }
            email: { type: string, format: email }
            isEmailVerified: { type: boolean }
        accessToken:
          { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        refreshToken:
          { type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, isEmailVerified]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        isEmailVerified: { type: boolean }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
    VerifyEmailResponse:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
    ChangePasswordResponse:
      type: object
      required: [message, accessToken, refreshToken]
      properties:
        message: { type: string, example: "Password updated" }
        accessToken: { type: string }
        refreshToken: { type: string }

    # Content Schemas
...
    ValidationErrorResponse:
      type: object
      required: [message, errors]
      properties:
        message: { type: string, example: "Validation error" }
        errors:
          type: object
          additionalProperties: true
          example:
            email: ["Invalid email format"]
            password: ["Password must be at least 8 characters"]
...
    UnauthorizedErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Invalid token" }
        error: { type: string, example: "Unauthorized" }
...
security: [] # default none; per-route add as needed
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: User signup with email verification
      description: |
        Creates a new user account and sends verification email.

        **Rate Limit:** 10 requests per minute per IP

        **Password Requirements:**
        - Minimum 8 characters
        - At least one uppercase letter
        - At least one lowercase letter
        - At least one special character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, cnic, password, province]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                email: { type: string, format: email }
                phone: { type: string, minLength: 5, maxLength: 30 }
                cnic:
                  {
                    type: string,
                    minLength: 5,
                    maxLength: 30,
                    description: "Pakistani CNIC",
                  }
                password:
                  {
                    type: string,
                    minLength: 8,
                    pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*[^A-Za-z0-9]).{8,}$",
                  }
                province:
                  {
                    type: string,
                    enum: [Punjab, Sindh, KP, Balochistan, ICT],
                  }
                language: { type: string, enum: [en, ur], default: en }
                city: { type: string, nullable: true }
                gender: { type: string, nullable: true }
                age: { type: integer, nullable: true }
                fatherName: { type: string, nullable: true }
                fatherCnic: { type: string, nullable: true }
                motherName: { type: string, nullable: true }
                motherCnic: { type: string, nullable: true }
                totalSiblings: { type: integer, default: 0 }
                brothers: { type: integer, default: 0 }
                sisters: { type: integer, default: 0 }
                timezone: { type: string, default: "UTC" }
            example:
              name: Ahmed Khan
              email: ahmed@example.com
              phone: "03001234567"
              cnic: "12345-1234567-1"
              password: "SecurePass@123"
              language: en
              province: Punjab
              city: Lahore
              gender: male
              age: 28
              fatherName: Khan Sahib
              fatherCnic: "11111-1111111-1"
              motherName: Begum Sahiba
              motherCnic: "22222-2222222-2"
              totalSiblings: 3
              brothers: 2
              sisters: 1
              timezone: Asia/Karachi
      responses:
        "201":
          description: User created successfully
          headers:
            X-RateLimit-Limit:
              { schema: { type: integer }, description: "Rate limit ceiling" }
            X-RateLimit-Remaining:
              { schema: { type: integer }, description: "Requests remaining" }
            X-RateLimit-Reset:
              {
                schema: { type: integer },
                description: "Unix timestamp when limit resets",
              }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignupResponse" }
        "400":
          description: Validation error or missing fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "409":
          description: Email or CNIC already exists
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
        "429":
          description: Too Many Requests (10 per minute)
          headers:
            Retry-After:
              {
                schema: { type: integer },
                description: "Seconds until retry allowed",
              }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: |
        Authenticate user and receive access tokens.

        **Rate Limit:** 10 requests per minute per IP

        **Requirements:**
        - Email must be verified
        - Account must not be deleted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 1 }
            example:
              email: ahmed@example.com
              password: "SecurePass@123"
      responses:
        "200":
          description: Login successful
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string, example: "Invalid email or password" }
        "403":
          description: Email not verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Email not verified" }
        "429":
          description: Too Many Requests (10 per minute)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Use refresh token to get new access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email address
      description: |
        Email verification via token sent to user's email.

        **Rate Limit:** 30 requests per hour per IP
        Returns `verified: false` if the token is invalid or expired.
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
          description: "Verification token from email (valid for 24 hours)"
      responses:
        "200":
          description: Email verification result
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyEmailResponse" }
        "400":
          description: Missing token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too Many Requests (30 per hour)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: |
        Sends password reset email to user.

        **Rate Limit:** 5 requests per minute per IP

        **Note:** Returns success even if email doesn't exist (security best practice)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
            example:
              email: ahmed@example.com
      responses:
        "200":
          description: Password reset email sent (if account exists)
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too Many Requests (5 per minute)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: |
        Reset password using token from email (valid for 1 hour).

        **Rate Limit:** 5 requests per minute per IP

        **Invalidates all existing tokens** for security.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword:
                  {
                    type: string,
                    minLength: 8,
                    pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*[^A-Za-z0-9]).{8,}$",
                  }
            example:
              token: "abc123def456..."
              newPassword: "NewSecure@Pass456"
      responses:
        "200":
          description: Password reset successful
          headers:
            X-RateLimit-Limit: { schema: { type: integer } }
            X-RateLimit-Remaining: { schema: { type: integer } }
            X-RateLimit-Reset: { schema: { type: integer } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Invalid/expired token or validation error
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - $ref: "#/components/schemas/Error"
        "429":
          description: Too Many Requests (5 per minute)
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RateLimitError" }
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password (authenticated)
      description: |
        Change password for authenticated user.

        **Returns new tokens** after password change.
        **Invalidates all previous tokens** for security.
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmPassword]
              properties:
                currentPassword: { type: string }
                newPassword:
                  {
                    type: string,
                    minLength: 8,
                    pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*[^A-Za-z0-9]).{8,}$",
                  }
                confirmPassword: { type: string }
            example:
              currentPassword: "SecurePass@123"
              newPassword: "NewSecure@Pass789"
              confirmPassword: "NewSecure@Pass789"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangePasswordResponse" }
        "400":
          description: Validation error or current password mismatch
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationErrorResponse"
                  - type: object
                    required: [message]
                    properties:
                      message: { type: string }
        "403":
          description: Safe mode enabled (no password changes allowed)
          content:
            application/json:
              schema:
                type: object
                required: [ok, message]
                properties:
                  ok: { type: boolean, example: false }
                  message: { type: string, example: "Safe mode" }

  # -------------------- USERS --------------------
...